tests/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = "rust-bcli tests"
authors = ["Andrew Toth <andrewstoth@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-timeout = "^2.1.0"
pyln-testing = { path = "../../lightning/contrib/pyln-testing/", develop = true }


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
